+++ src/index.ts
@@ -2,7 +2,10 @@
 export { SoundCloudClient } from "./client/SoundCloudClient.js";
 export type { SoundCloudClientConfig, TokenOption } from "./client/SoundCloudClient.js";
 export { scFetch, scFetchUrl } from "./client/http.js";
-export type { RequestOptions } from "./client/http.js";
+export type { RequestOptions, RetryConfig } from "./client/http.js";
+
+// Errors
+export { SoundCloudError } from "./errors.js";
 export { paginate, paginateItems, fetchAll } from "./client/paginate.js";
 
 // Types